chore: l10n_countries breaking change and renamed country name as "localizedName"
New file:
* `simple_input_page_country_helper.dart`: Implementation for "Countries" of an [AbstractSimpleInputPageHelper]. Moved from `simple_input_page_helpers.dart`

Impacted files:
* `country.dart`: complying with the l10n_countries breaking change; renamed poorly named `name` (for an `enum`) to `localizedName`
* `country_selector.dart` : now using explicitly `localizedName`; minor refactoring
* `country_selector_provider.dart` x 2: minor refactoring
* `country_selector_tile.dart`: : now using explicitly `localizedName`; minor refactoring
* `edit_product_page.dart`: minor refactoring
* `GeneratedPluginRegistrant.swift`: wtf
* `knowledge_panel_action_card.dart`: minor refactoring
* `price_to_oxf.dart`: minor refactoring
* `product_query_page.dart`: now using explicitly `localizedName`
* `pubspec.lock`: wtf
* `pubspec.yaml`: upgraded `l10n_countries` to `2.0.2`
* `simple_input_page_helpers.dart`: moved country to a specific file; minor refactoring
