+++ packages/smooth_app/lib/pages/preferences_v2/tiles/app_settings_tiles/country_selector/country_selector_tile.dart
@@ -77,7 +77,8 @@ class CountrySelectorTile extends PreferenceTile {
                       spacing: VERY_SMALL_SPACE,
                       children: <Widget>[
                         Text(
-                          country?.name ?? AppLocalizations.of(context).loading,
+                          country?.localizedName ??
+                              AppLocalizations.of(context).loading,
                         ),
                         if (country != null)
                           SizedBox(
@@ -239,7 +240,7 @@ class _CountrySelectorScreen extends StatelessWidget {
                 Expanded(
                   flex: 7,
                   child: TextHighlighter(
-                    text: country.name,
+                    text: country.localizedName,
                     filter: filter,
                     textStyle: const TextStyle(fontWeight: FontWeight.w600),
                   ),
@@ -253,31 +254,12 @@ class _CountrySelectorScreen extends StatelessWidget {
             Country? selectedItem,
             Country? selectedItemOverride,
             String filter,
-          ) => _filterCountries(
+          ) => OpenFoodFactsCountryLocalization.filterCountries(
             list,
             selectedItem,
             selectedItemOverride,
             filter,
           ),
-    );
-  }
-
-  Iterable<Country> _filterCountries(
-    List<Country> countries,
-    Country? userCountry,
-    Country? selectedCountry,
-    String? filter,
-  ) {
-    if (filter == null || filter.isEmpty) {
-      return countries;
-    }
-
-    return countries.where(
-      (Country country) =>
-          country == userCountry ||
-          country == selectedCountry ||
-          country.name.toLowerCase().contains(filter.toLowerCase()) ||
-          country.offTag.toLowerCase().contains(filter.toLowerCase()),
     );
   }
 }
