+++ packages/smooth_app/pubspec.yaml
@@ -29,7 +29,7 @@ dependencies:
   http: 1.6.0
   http_parser: 4.1.2
   image_picker: 1.2.1
-  l10n_countries: 1.3.1
+  l10n_countries: 2.0.2
   latlong2: 0.9.1
   matomo_tracker: 6.1.0
   package_info_plus: 9.0.1
